Support jruby - the open4 gem isn't compatible, use `IO.popen4` with jruby
The `open4` gem isn't compatible with jruby because jruby doesn't allow forking.
To allow Scmd to be used with jruby, we can use `IO.popen4`, which provides the
same functionality as the `open4` gem. This commit modifies the command class to
switch `Open4` usage with `IO` when the platform is java.