update big data capture test to allow more time
This test was randomly failing b/c it would timeout.  This ups
the allowed time to minimize the chance of a timeout happening.

In this case it is ok b/c the timeout isn't being tested.  The
streaming of a large amount of output is being tested.  We want
to ensure there is enough time given to stream all that data.

/cc @jcredding