Update posix-spawn dependency
This updates the posix-spawn dependency to the latest version and
also switches back to allowing any patch level version updates.

This was originally locked down in c3206e6 because the latest
version of posix-spawn was not compatible with ruby 1.8.7. They
fixed the 1.8.7 errors in v0.3.10.

As an aside, this also fixes an issue with ruby 2.2 we were
experiencing in other gems. In Qs, its benchmark script relies on
creating a unix pipe, spinning up a child process that runs jobs
and passing the writer file descriptor to that child process. In
ruby 1.9 they added `Process.spawn` and by default it closes all
file descriptors to avoid "leaking" file descriptors which can
apparently keep files from being closed properly. posix-spawn tries
to be compatible with ruby so it does the same thing. So in ruby
2.2, qs benchmark script won't work unless it redirects file
descriptors. There was a bug with redirecting the file descriptors
with posix-spawn in v0.3.8 though. By updating the posix-spawn
version, the Qs benchmark script can work as expected in ruby 2.2.