Allow passing options to posix-spawn
This allows passing options to posix-spawn which allows extra
behavior for the child process that is spawned.

This changes the command interface to take a command string and
opts hash instead of an env hash. The env hash can now be specified
using the `:env` key in the opts hash. This allows passing an
`:options` key in the opts hash for specifying the posix spawn
options.

This is needed for ruby 1.9+. It closes file descriptors when a
child process is spawned. This specifically caused Qs benchmark
script to not work anymore because it used an IO pipe to
communicate between the parent and child process. With this change
the Qs benchmark script can specify the file descriptors it wants
to redirect and the script works as it previously did.

This also updates the README documentation for using the new
options. This includes a section showing how environment variables
can be used which should have been added when environment variable
support was added.