add "test mode" for Scmd
Test mode turns on a bunch testing utilities for Scmd.  This means
that command spies will be used in place of "live" commands, each
time a command is run or started will be logged in a collection and
option-specific spies can be added and used to "stub" spies with
specific attributes in specific contexts.

In non-test-mode, everything behaves as before: `.new` returns
"live" commands and the test mode API all raises `NoMethodError`s.
To turn on test mode, set and env var: `ENV['SCMD_TEST_MODE'] = '1'`.

Note: this also updates the command spy to have non-nil default
values for exit status and pid.  The goal here is to default with
realistic values that can be overridden and also so that commands
are successful by default.  The original defaults were a big
nuisance when used in application testing.