Hotfix: `CommandSpy`, don't add calls to `Scmd.calls` if not in test mode
This updates the `CommandSpy` to not add calls to `Scmd.calls` if
its not in test mode. This allows the `CommandSpy` to be used even
if scmd is not in test mode. This is specifically needed for dk
because it wants to use the `CommandSpy` directly without putting
all of scmd in test mode.

The `CommandSpy` now checks if the test mode env var is set before
it tries pushing a call onto `Scmd.calls`. This avoids `Scmd.calls`
raising a `NoMethodError` whenever a `CommandSpy` is run.