Hotfix: Add comment to Gemfile explaining how to release new versions
This updates the Gemfile to add a comment explaining how to release
scmd. Since we use `ggem` to release new gem versions and scmd is
a dependency of `ggem` the normal process doesn't work. This is
because `ggem` will load the installed version of scmd instead of
loading the scmd files. To fix this, we need to run:
`bundle exec ggem` so bundler will make it require the scmd files
instead of the installed version. To do this, `ggem` has to be
added to the gemfile. Also since adding ggem to the gemfile will
mean there is an uncommited file, the `-f` flag has to be used to
force releasing the gem. So this adds a commented out `ggem` to the
gemfile and adds a comment explaining to do a forced release.